requirements.txt
amqp==5.2.0
asgiref==3.4.1
async-timeout==4.0.3
billiard==4.2.0
black==22.12.0
boto3==1.26.145
botocore==1.29.145
cachetools==5.3.1
celery==5.3.6
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
chessql==2.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==3.2.7
django-cors-headers==3.10.0
django-crontab==0.7.1
django-environ==0.9.0
django-filter==2.4.0
django-health-check==3.17.0
django-ses==3.5.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.9.0
flake8==6.0.0
freetype-py==2.3.0
future==0.18.2
gunicorn==20.1.0
identify==2.5.13
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.4
libcst==0.4.10
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
openapi-codec==1.3.2
packaging==21.3
pandas==1.5.3
pathspec==0.11.0
pgn-read==2.2
Pillow==9.5.0
platformdirs==2.6.2
pre-commit==2.21.0
prompt-toolkit==3.0.42
psycopg2-binary==2.9.5
pycairo==1.24.0
pycodestyle==2.10.0
pydub==0.25.1
pyflakes==3.0.1
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==5.0.1
remove-print-statements==0.5.2
reportlab==4.0.4
requests==2.27.1
rlPyCairo==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
SpeechRecognition==3.10.0
sqlparse==0.4.2
tomli==2.0.1
typing-inspect==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.1.0
virtualenv==20.17.1
wcwidth==0.2.12
